Normalize indents, CSS
