+++ app/assets/stylesheets/shared/_app.scss
@@ -5,7 +5,7 @@ body {
   scrollbar-gutter: stable;
 }
 
-.hidden_link {
+.hidden-link {
   text-decoration: none;
 
   &:hover, &:focus {
