+++ app/javascript/controllers/form/visibility_controller.js
@@ -1,14 +1,14 @@
 import { Controller } from "@hotwired/stimulus"
 
 export default class extends Controller {
-    static targets = ["checkbox", "element"]
+  static targets = ["checkbox", "element"]
 
-    connect() {
-        this.toggle()
-    }
+  connect() {
+    this.toggle()
+  }
 
-    toggle() {
-        let visible = this.checkboxTarget.checked;
-        this.elementTargets.forEach((element) => element.style.display = visible ? "block" : "none");
-    }
+  toggle() {
+    let visible = this.checkboxTarget.checked;
+    this.elementTargets.forEach((element) => element.style.display = visible ? "block" : "none");
+  }
 }
