+++ app/views/admin/_footer.html.erb
@@ -3,7 +3,7 @@
     High School Hackathons Dashboard
     <br/>
     Made with ❤️ by
-    <%= link_to "Hack Club", "https://hackclub.com", class: "hidden_link" %>
+    <%= link_to "Hack Club", "https://hackclub.com", class: "hidden-link" %>
   </p>
 
 </footer>
