+++ app/views/admin/header/nav/_link.html.erb
@@ -1,5 +1,5 @@
 <% selected = @nav_active_item == link %>
 
-<%= link_to link, class: "admin__nav__link#{"-active" if selected }" do %>
+<%= link_to link, class: "admin__nav__link#{"--active" if selected }" do %>
   <%= title %>
 <% end %>
